Refactor package structure for materialized view classes
